implemented NN and the evolutionary algorithm. (missing fitness function)
Added the needed algorithms in general (not tested yet).
A fitness function is missing. -> so the network can be trained.
Actual execution of actions during trainnig is not impleemnted yet(i will take care about that.)
Calculation of the Sensor values has to be implemented.
Actual movement should be visible.